Fix bug when using Elixir 1.6 (#15)
Elixir 1.6 changed the behavior of Regex.split:

> [Regex] Return a leading empty space when splitting on empty pattern. This makes the split operation consistent with the other operations in the Regex module

The fix is to reject "" strings which returns the normal behavior. Tests passed in Elixir < 1.6, but did not pass in Elixir >= 1.6 until this fix.

Elixir 1.6 formats the mix.lock file differently. The versions don't actually change.
Add .tool-versions file for local development to enforce the same Elixir and Erlang versions.