Support releases (#18)
Resolves #17 

This will result in a breaking change for other users, since it introduces a required configuration. 

Another change is the default folder, however using `Application.app_dir` should work for Phoenix 1.2, 1.3, and 1.4 since Phoenix symlinks the `priv` folder inside the `_build` folder (which is where `Application.app_dir` would resolve within), but I have not tested this with Phoenix 1.2 to be sure of that.